[dw1000] Use a 32-bit multiplier for frequency adjustments
A 24-bit multiplier provides a frequency adjustment maximum resolution
of around 60ppb.  This is not sufficient to obtain PTP clock
synchronisation within a few nanoseconds.

Upgrade to a 32-bit multiplier by separating the actively counting
upper 31 bits of the 40-bit timestamp from the derived lower 9 bits.
This allows for a frequency adjustment maximum resolution of around
0.25ppb, which in turn allows PTP to synchronise to within single
digits of nanoseconds.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>