[dw1000] Calculate link quality as fp_ampl2 / std_noise
Calculate a link quality indicator as fp_ampl2 divided by std_noise
(limited to a maximum of 0xff) and report as the lqi value to
ieee802154_rx_irqsafe().  The IEEE 802.15.4 core stores this value in
mac_cb(skb)->lqi and then ignores it; there appears to be no way to
retrieve this value from userspace.

Provide a configurable link quality indicator threshold below which
the receive timestamp will be ignored.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>