[dw1000] Do not reset receiver when an overrun occurs
Experiments show that the receiver reset process is massively
disruptive to both transmit and receive datapaths, and prone to a
failure mode in which the RXENAB command is simply ignored, resulting
in a completely dead receiver.

In contrast, simply toggling HRBPT seems to work reliably.  The only
subtlety is that we need to treat an RXOVRR interrupt in the same way
as an RXDFR interrupt, since RXDFR will not be asserted in the event
of an overrun.

Remove the complexity of the receiver reset mechanism, and simply
discard any packets for which an overrun has occurred.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>