script: Check in clockrate.pl to measure DW1000 clock against system clock
We have been using the average of a bunch of DWM1000 modules to find the
crystal trim. This is obviously without reference to an absolute timing
source, so it would be useful to be able to calibrate to an absolute
reference.

This script is a proof of concept to calibrate the DWM1000 crystal against
the system clock. This is obviously only useful if the system clock is
accurate, so use ntpd to ensure this. It may be necessary to run this script
for a long time to achieve consistent results.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>