firmware: full accelerometer support
The accelerometer is now probed on v1 and v2 pins, and is now used in
tagipv6 functionality to set a variable blink interval based on motion.
Configuration parameters exist to set the sensitivity of the acceloremeter
while in sniff mode, as well as to specify a time for which the tag will
remain in active mode after the last detected motion.

In its present form, the code will cause the tag to enter active mode
approximately once every 1.52 days, when the timer wraps around. This
should be fixed eventually.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>