case: Add v2a prototype case.
This is designed for a single M2 countersunk bolt with a hex nut. A 9mm
bolt will be slightly too long and will stick out, an 8mm bolt will probably
work but won't engage with the entire thickness of the nut.

There's an absence of support over the battery area. This may need improving.
Also, the single screw design may be insufficient, depending on the
flexibility of the material.

This is mostly designed to be injection mouldable, although there are still
a few areas where corners are potentially too sharp for that process. Once
we're happy with the design, a review by an injection moulding expert would
be a very good idea.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>