firmware: Measure battery voltage with ADC and report in blink packets
Also, check in file that was missed off a previous commit.

Note that at present we're just using a linear estimator between 3.6
volts and 4.1 volts for state of charge. This isn't particularly
accurate, but gets us going until we revisit this to come up with
a more accurate state of charge estimator.

Also, the battery voltage is filtered with an exponential weighted
average filter to keep it stable. The tag needs to wake up about 16
times after a significant change to the battery voltage (such as
fast charging while the tag is idle) to start reporting an accurate
state.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>