firmware: Time the rx window from the start, not from the last restart
We need to use a timer on the microcontroller for this, as the radio is
not capable of using its rx timeout functionality in this way. So we
just disable the rx timeout on the radio and take care of it ourselves.

This is non-trivial because there is no fast accurate clock on the
microcontroller. The system is clocked from an RC oscillator, which
could be off by varying amounts depending on the environment.

So, we estimate the time based on the nominal frequency and then
dynamically adjust the time based on the actual time of the previous
window, measured by reading timestamps from the dw1000.

Currently there's no filtering on this, but it would make sense to
refine it to slowly adjust the timing to drift towards the right answer,
thus eliminating instabilities and jitter from the measurement.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>