firmware: implement OTA write
This implements OTA write. There must be sufficient space
available in config for the entire write to complete, otherwise
it is rejected. An extra deletion step is done prior to the writes
so that the config won't temporarily exceed capacity, but only if
the data needs changing.

A few bugs in the config module have been fixed along the way.

Also some new cli commands:

* free - returns the free space available in the config area
* dump - dumps the internal data structure of the config area

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>