circuit: Add UART-USB level shifting and reduce charge current
We need to connect the UART to the USB data pins. As these will present
5 volts, we need to add level shifting to protect the 3.3 volt mcu.

We will need to enable the internal pullup on the UART Rx line.
The host cable we are proposing to use already has a pullup on its
Rx line.

While we're here, reduce the charge current to just under 200mA in
order to meet the 0.5C charge current specification on the 400mAh
batteries that we're using.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>