firmware: hack to use a tail v1 tag to unbrick another tag
If you have bricked your tail tag, this is the method of last
resort to recover it. It uses a tail v1 tag (which has spare
GPIO pins brought out to a header) in order to implement just
enough of SWD to connect to the Authentication Access Port
(available for a very short window after reset) and perform
a mass erase of the flash.

Connect the following to the GPIO header on a tail v1 tag:

Pin 1 - SWCLK
Pin 2 - SWDIO
Pin 3 - RESET
Pin 15/16 - GND

Connect these via jumper wires and breakouts to the DEBUG connector
on the target tag.

Then send the "unbrick" command to the tail v1 tag.

You should see something like this:

IDCODE: BC11477
IDR: 4770031
AAPIDR: 16E60001
Status: 0

If you see that, your target tag should now be erased and unbricked.
Everything will be erased, including the tag config.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>