bootloader: Enter bootloader only if Ctrl-C is received
The UART is enabled earlier, in an Rx only configuration. The chip
wakes up from time to time to poll the UART - if a Ctrl-C is received
then we enter the bootloader. If we time out, then we see if we can
boot the firmware. If we still can't boot the firmware, we launch
the bootloader anyway.

Signed-off-by: Gavan Fantom <gavan@coolfactor.org>