firmware: Listen periodically if listen_period is set
A new config key, listen_period, sets the number of seconds
after which, if a receive window has not been opened, the tag
will listen even if there is no other reason to. This ensures
that it is possible to reconfigure a tag with OTA config if
it is configured to do one way ranging only.

Note that if the key is not present, a receive window will
only ever be opened if it is needed for other operations.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>