scripts: send.sh: Don't echo what the tag sends back to it
For some reason linux has been echoing data from the tag back
to the tag on the UART. This doesn't seem to have been the case
when developing or demonstrating this script previously, but
it has shown to be a problem on at least two people's laptops today.

So we explicitly turn off echo.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>