prodtest2a: design proper PCB
This integrates the relay functionality as well as the tester
onto a single PCB, designed to fit within an instrument enclosure.

Custom front and rear panels will need to be manufactured, probably
by 3D printing, to match this layout.

It is intended that the fixed test hardware will be connected to the
back panel, and that the DUT (via the jig) will be connected to the
front panel.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>