firmware: Add supervisor to check for failed blink events
We add a 100ms timer starting at each blink event. If the
radio is still active when this times out, it is assumed
that something has gone wrong. We capture the DW1000 state
register to aid debugging, and record the number of times
that this condition has occurred. These values can be
read with the "supervisor" command. If the count is non-zero
then we will also start transmitting an IE containing
these values.

When we do hit this condition, we also send TRXOFF to the
radio and set radio_active to false, in the hope that
this will be sufficient to unblock the radio and allow
the firmware to continue at the next blink event.

It may be necessary to take harsher measures in the future,
but for now we hope this will be enough.

Also, fix a bug in the scheduling of early wakeups in the
presence of multiple timers and the cancellation of a timer,
and remove some extraneous debug prints.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>