firmware: transmit cookies
Transmit 'cookies' (an IV precursor) if sufficient entropy is available on the
tag and tx_cookie is set to true in the tag config.

This patch includes explicit alignment of byte arrays that are cast to
uint32_t* in emlib. The CPU is configured to halt on unaligned access.

This also contains DRBG fixes (calling AES_ECB128 with lengths of 16 bytes).

This also contains code to calculate an IV from the cookie, using the
appropriate derived key. This has been verified against the Python server
implementation.

Signed-off-by: Luke Barone-Adesi <luke.barone-adesi@unipart.io>