measurement: Script to calculate location in the room
Given a Z coordinate (which you must measure) and
a number of distances from existing anchors (read from
config.json in the same directory) this script will
attempt to calculate the X and Y coordinates of
the measured object.

Enter a number in the same units of measurement as used
in config.json to represent a distance from the named
anchor to the unknown location, or a blank line if no
such measurement is available. Once all anchors in
config.json have been queried, a calculation will be
performed.

Signed-off-by: Gavan Fantom <gavan.fantom@unipart.io>